Consolidate build test names and description
Update build test image names to be more generic and not
bound to a specific version of a distribution. As the tests
are usually based on rolling releases of distros the name
of the test image should be generic. Also adapted the
specification of the test images to describe the focus
of the test if not generic.