Add support for rd.root.overlay.readonly
Allow to use an overlayroot system with read-only feature.
If kernelcmdline="rd.root.overlay.readonly" is set in combination
with overlayroot="true" the overlayfs uses a tmpfs to store
new data temporary as long as the system runs. On reboot the
newly written data is lost and the system is back to its
factory state.