Fixed device_array structure in get_selected_disk
In reference to Issue #880 a bug was introduced that broke
the contents of the device_array which causes issues on
installations with two or more attached disks. The change
in the mentioned PR reduced the tuple for each disk
from 3 elements to 2 elements. Therefore the loop that
iterates over the disk tuples via modulo 3 was broken.
This commit fixes the modulo operation to correctly
parse the disk_list. Fixes #1588