Rename image build tests
To use the image builds in openQA they have to have a unique
name such that it cannot happen that a cached version of an
image in openQA is used. The current names matched openQA
cached images e.g openSUSE-Tumbleweed and in addition different
image build tests used the same name. This commit uses the
name of the image as it is organized in its directory structure
prepending "kiwi-" to be unique in openQA when it fetches
the image. This is realted to Issue #1555