Added support for nbd and aoe root overlay
The kiwi-overlay dracut module can also be used as standalone
module that is not connected to a disk image. In this case
it's needed to specify the location for the root filesystem
and optionally the device to write data (default is ram space).
This commit adds the opportunity to specify a nbd/aoe location
for the root filesystem on the kernel cmdline like in the
following examples:

    root=overlay:nbd=nbd0:192.168.100.42:exportname
    root=overlay:aoe=e0.1

An optional write space, if it should not be ram space, can be
provided through the rd.root.overlay.write option on the kernel
cmdline. This Fixes: OSInside/kiwi-descriptions#78