Avoid using generators in pre-mount hooks
This commit deletes the generator that was creating the sysroot.mount
unit for ramdisk deployments. Generators, specially the sysroot.mount is
expected to be created on very early stages of the boot procedure as
this has impact on relevant targets such as initrd-root-fs.target, which
does not depend on sysroot.mount if the unit is not there.

In ramdisk deployments some data is known on pre-mount stage as
as it is downloaded from the PXE server. At this stage it is not safe to
generate a sysroot.mount unit that depends on initrd-root-fs.target as
the target is close to finalize or even finalized already and could
potentially skip sysroot.mount exection.

Instead we include a mount hook which is only executed on ramdisk
deployments that simply runs the mount command to mount /sysroot.

This fixes bsc#1178670