Update packaging spec
Get rid of spec file Recommends because this is not implemented
consistently across all distro rpm versions. In addition the
recommended software components only makes sense if used in an
image build context. To create context a collection of

    kiwi-host-setup-for-CONTEXT

sub packages are added which pulls in the context required
software components. This Fixes #1503