Delete obsolete ddb.adapterType patching
When building a vmdk image with pvscsi as adapter type, kiwi
implicitly changed the adapter_type from pvscsi to lsilogic
because the pvscsi driver was missing. The real solution to
this problem is to include the now existing vmw_pvscsi kernel
driver into the initrd with a custom root overlay file
etc/dracut.conf.d/pvscsi.conf. This Fixes bsc#1180539