Added strong typing for the following API methods
* kiwi/system/prepare.py
* kiwi/system/profile.py
* kiwi/system/setup.py

The changes here also lead to a small refactoring for the handling
of the package manager. In my opinion it doesn't make sense to allow
a None type package manager from the stateful XML instance. As without
any package manager nothing can be done. As it also turns into an
issue for the PackageManager API which does not allow for an empty
value here I thought it's better to come up with a default package
manager (set to dnf) if no one is explicitly specified

This references Issue #1644