Fixed creation of custom grub image
If kiwi can't find the distribution provided grub image(s) it
creates them using the respective grub[X]-mkimage tool. However
the tool was called on the build host which could cause an
inconsistency on the used module path. Grub is not packaged
consistently across the distributions and also the provided
modules comes in different versions and patch sets. Therefore
it's required that kiwi calls the grub mkimage tool as chrooted
operation inside of the tree that provides the target image.
As consequence of the change it's required that the image root
tree provides the grub mkimage tool. In addition to the change
we now also log the output from the grub mkimage call in the
kiwi build logfile. This Fixes #1254