Added option to allow reading obs project repos
If kiwi is used in OBS many people setup the repositories
in the OBS project metadata. In the KIWI image description
only a placeholder called obsrepositories needs to be added
to indicate that OBS is managing the repositories. This is
nice but prevents people from simply building this image
outside of OBS. This commit adds a new caller option:

    --set-obs-repos=<user,password,project,name,profile,arch,repo>

which allows to read and use the list of repositories as OBS
would use it but in a local build process. KIWI will send a _buildinfo
request to the OBS API and reads the provided result. For this
call to work the user needs to provide its obs account information.
The account data is only used in a HTTPBasicAuth request and
not stored or used in any other way.