Added option to allow building from OBS project
This commit implements a new build option --obs-image which
allows to build an image locally with the project and repo
definitions hosted in the open buildservice. The command
reads the image description from obs and checks it out locally.
Published repositories managed in the project configuration
on obs will be auto-expanded to become locally usable.
Project definitions based on external source services are
not supported yet.