Fixed ramdisk deployments
The hook script to mount the ramdisk to /sysroot was not called.
The script was installed to /lib/dracut/hooks/mount/30-kiwi-mount-ramdisk.sh
but was not called by dracut. The mount hooks are only called if NO
root= information is present in the cmdline. For ramdisk deployments
from an install ISO there is a root= parameter and this the concept
to put the mount code as a mount hook in dracut is broken by design.
This commit simply calls the code when needed and avoid the hook
system at dracut in this case