Fixed ramdisk deployment from install ISO
The dracut mount hook is only called if NO root= information
is provided from the cmdline. In a PXE deployment this is the
case but in an install ISO deployment not. This commit re-adds
the sysroot mount generator approach but only applies on the
condition rd.kiwi.ramdisk is set and root= is specified. Only
in this case we know config.bootoptions is present in the
initrd and will not be provided from an outside service