Move fedora integration tests to stable release
Instead of building against the rolling release (rawhide)
we build the integration tests for non suse tests against
the stable release of distributions. The reason is except
for suse we don't have influence and resources on the
distribution development and sometimes facing the problem
that we can only sit and pray for a self healing of the
rolling distro releases. From a kiwi release perspective
this blocks us too often