Delete support for generic iso:// source type
The generic iso:// media type mounts the given iso file and expect
its root to provide a repository that can be used 1:1 with a package
manager. This concept is broken since some time and it can't be fixed
in a generic way. All distribution media comes with a certain layout
and basically needs extra handling to become fully usable as
repository. The current implementation of the iso type which simply
mounts the iso and expects its root to be a known repo is not useful.
Therefore the support for it will be decommissioned. Instead we
will provide a documentation chapter that documents how to incorporate
distro ISO media for building images.