Fail non-XML descriptions if anymarkup not available
If the anymarkup Python module is not available, and a non-XML format
description file has been provided, then fail with an error indicating
that non-XML format config files require anymarkup.

Adds a new exception, KiwiConfigFileFormatNotSupported, to be raised
for this scenario.

Added test_apply_xslt_stylesheets_nonXML() to the markup base unit
tests to test this scenario.

Fixes #1702