Fixed LVM root volume inclusion in fstab
The root(/) volume in a LVM setup was included as UUID device
path. This is inconsistent compared to the non root volumes
which are included by their LVM representation:

    /dev/VGROUP_NAME/VOLUME_NAME

This commit makes sure the root volume is included by its
persistent LVM device representation too. This is related
to Issue #1747