Fixed integration tests requesting grub
For some reason two integration tests requests the "grub" and
the "grub2" package. On TW there is now no longer a provider of
grub and thus the image build became unresolvable. There is
only grub2 for quite some time, so I deleted this package
setting from the tests in question