Change zypper download mode to in-advance
In relation to upcoming zypper changes e.g to make use of librpm on
single transaction operations there is the possibility that file
triggers start being used. To ensure zypper behaves consistently
DownloadInAdvance mode should be used, this way the transaction
happens as a whole and with the upcoming zypper changes zypper
will still be capable to handle the file triggers.
This Fixes #1789