Fixed spec file on Debian tool settings
debootstrap should always come with dpkg because we don't
want to handle the architecture names used on Debian when
dpkg knows them better than we do. since debootstrap itself
considers the possibility of being called without dpkg on
the system we generate the dependency by a spec change
here. This Fixes #1778