Fixed live network setup to be generic
In dracut the network setup comes with different models
providing a different set of functions. The ifup method as
used in the live iso dracut module is only available with the
network-legacy mode and fails with network-wicked. This commit
uses a dracut conf file in /etc/cmdline.d which uses the dracut
network interface parameters instead of calling module specific
methods. This Fixes #1802