Update documentation qemu calls
We use the kiwi integration tests as base for the documentation
example images now. The integration tests are all configured
to set the console to serial. Thus the docs should explain
the qemu call for test runs using the -serial stdio option
to make sure the console information is displayed to the
user