Add log information on grub search
There is a method in kiwi which searches for grub files.
As grub is packaged differently within the distributions
a dynamic lookup is needed. However, the result and where
kiwi looked it up was not part of the log file. In terms
of issues like the one from Issue #1754 it would be very
handy to know about this information. Thus this commit
adds debug information to the log file regarding what
grub files are searched and where and if found