Update Ubuntu integration test for system settings
In Debian based distributions the kiwi built in way
to setup locale, keyboard and timezone via systemd tools
does not work because not(yet) provided by the distribution.
This commit adds a reference implementation in the Ubuntu
integration test to demonstrate how the settings given in
the kiwi image description needs to be handled to make
them effective in the later image. This Fixes #1787