Do not return default stdout if it is no raising on failure
This commit prevents the use of a default stdout and stderr in case
return code reports errors and it is not raising an exception.

If we are not raising an exception there is no specific need to
artificially append some stdout and stderr default message, we just
behave as if there was no error.