Fixed creating grub bios module
If no prebuilt grub bios module was found, kiwi creates one.
In this case kiwi searches for the grub modules and runs
the grub mkimage tool. The search for the modules for the
bios module used the host system (/) grub and that fails if
the host has packaged grub differently than the image target.
This fix moves the lookup into the image root directory
which is the correct place to lookup the grub data