Start moving CI test from gitlab to actions
Moving the linter and unit tests to github workflows and
out of the gitlab CI system has the advantage that pull
request from forked repos will run the tests. In the long
run I think we should move away completely from gitlab CI
and use github actions as this will reduce the number of
external services used in the kiwi project. This commit
starts with the most simple parts: unit and type tests,
doc rendering and code style