Start moving CI test from gitlab to actions
It would be imho good to use the github CI system if the
code is also on github. The reason for gitlab at the time
was a missing CI implementation for github. However, this is
no longer true and moving could simplify and cleanup the
CI situation for the KIWI project. This commit starts with
the most simple parts: unit and type tests, doc rendering
and code style