Don't modify main sshd_config file via baseVagrantSetup
In Tumbleweed we switched from shipping sshd's config file in /etc to /usr/etc,
but baseVagrantSetup was not respecting that.

So instead of modifying the main config file, we instead create a custom file in
/etc/ssh/sshd_config.d/ with our custom setting which is much more robust
accross distributions. However, if /etc/ssh/sshd_config.d/ does not exist, then
we fall back to adding our settings to /etc/ssh/sshd_config.