Refactor recovery archive creation
The creation of the recovery archive was based on a tempfile
which then gets moved to the final target file. This is bad
because if causes the tempfile cleanup to raise an exception
unless you specify delete=False. Creating a tempfile with
delete=False however, negates the reason for creating self
managed temporary data. Thus this code was refactored to
create the recovery archive file at its final destination
from the beginning on