Allow test-image-luks to be build locally too
In suse there are many programs replaced by the busybox
alternative and also preferred by the package manager.
However there are also packages like the kernel in suse
which wants gawk and not busybox-gawk. In kiwi to build
images there are two installation phases, the bootstrap
phase which installs mandatory packages to allow chroot
operations and the image phase which installs the rest
of the requested packages as chroot operation. In this
two pass step the package manager is called twice to
resolve dependencies and that causes the issue with the
preferred busybox packages which gets pulled in in the
first phase. To prevent this there are only two options:

1. Explicitly list the non busybox packages in the
   bootstrap phase

2. Prevent the image phase and put all packages to
   the bootstrap phase such there is only one package
   manager call for dependency resolution

This commit implementes option 1. because having all
packages in the bootstrap phase means that the host
packagemanager resolves and if the host packagemanager
is somehow not compatible with the image target this
leads to other weird issues which are not obvious and
hard to debug.