Update test for image without initrd and loader
The systemd-remount-fs.service expects devices to be present
in their UUID representation. When booting without an initrd
there is no udev and the root= device must be provided with
the kernel exposed unix device name. In addition there is no
UUID device representation for virtio devices which are used
together with the small kvmsmall kernel of this example.
Since booting of this appliance is done via:

    $ kvm --kernel (kernel-kvmsmall) \
        -drive file=kiwi-test-image-embedded.raw,if=virtio \
        -append "root=/dev/vda1 console=ttyS0 rw"

The kernel rootfs is mounted in rw mode anyway. Thus the systemd
remount service is not useful in this case and switched off