Fixed nasty bug in ramdisk generator
The ramdisk unit generator reads the config.bootoptions
file and extracts the root_uuid from it. This is done
with a very simple shell read using a space as separator.
However the last element is never read by that code.
As long as there was yet another kernel cmdline option
after the root= option this bug was not an issue. But
as soon as the root= option is last in the list it will
not be read and the generator exits prior creating the
sysroot.mount target. This commit fixes it in a way
that it makes sure there is always a space at the end
of the config.bootoptions file